fix meas value shorts

+++ src/IEC60870-5-104-simulator.Infrastructure/InformationObjectTemplateMethod.cs
@@ -72,8 +72,8 @@ namespace IEC60870_5_104_simulator.Infrastructure
             return type switch
             {
                 Iec104DataTypes.M_ME_NC_1 => new MeasuredValueShort(objectAddress, (float)value.GetValue(), new QualityDescriptor()),
-                Iec104DataTypes.M_ME_TC_1 => new MeasuredValueShortWithCP24Time2a(objectAddress, (int)value.GetValue(), new QualityDescriptor(), GetCP24Now()),
-                Iec104DataTypes.M_ME_TF_1 => new MeasuredValueShortWithCP56Time2a(objectAddress, (int)value.GetValue(), new QualityDescriptor(), GetCP56Now()),
+                Iec104DataTypes.M_ME_TC_1 => new MeasuredValueShortWithCP24Time2a(objectAddress, (float)value.GetValue(), new QualityDescriptor(), GetCP24Now()),
+                Iec104DataTypes.M_ME_TF_1 => new MeasuredValueShortWithCP56Time2a(objectAddress, (float)value.GetValue(), new QualityDescriptor(), GetCP56Now()),
                 _ => throw new NotImplementedException($"no {nameof(MeasuredValueShort)} for this type {type}"),
             };
         }
@@ -83,8 +83,8 @@ namespace IEC60870_5_104_simulator.Infrastructure
             return type switch
             {
                 Iec104DataTypes.M_ME_NA_1 => new MeasuredValueNormalized(objectAddress, (float)value.GetValue(), new QualityDescriptor()),
-                Iec104DataTypes.M_ME_TA_1 => new MeasuredValueNormalizedWithCP24Time2a(objectAddress, (int)value.GetValue(), new QualityDescriptor(), GetCP24Now()),
-                Iec104DataTypes.M_ME_ND_1 => new MeasuredValueNormalizedWithCP56Time2a(objectAddress, (int)value.GetValue(), new QualityDescriptor(), GetCP56Now()),
+                Iec104DataTypes.M_ME_TA_1 => new MeasuredValueNormalizedWithCP24Time2a(objectAddress, (float)value.GetValue(), new QualityDescriptor(), GetCP24Now()),
+                Iec104DataTypes.M_ME_ND_1 => new MeasuredValueNormalizedWithCP56Time2a(objectAddress, (float)value.GetValue(), new QualityDescriptor(), GetCP56Now()),
                 _ => throw new NotImplementedException($"no {nameof(MeasuredValueNormalized)} for this type {type}"),
             };
         }
